Refactored SQL statement to Prepared statement and added debug mode
The method DatabaseProvider::allItemsFromTable contained a possible
fault because a Statement object was being used that was built with a
user built input string. To prevent a SQL Injection, a prepared
statement is now used.
For debugging this software, I added immutable var DEBUG in Main.java
This can be removed in master branch when pushing the software in
production.